Moved validation_helpers.hpp from kernel_lib to kernel_helper_functions
